requirements.txt
Flask
slack_sdk
python-dotenv
newrelic
